New Changes -- Change Path For Home Page and new signup/login. Better ui

+++ public/js/logout-linker.js
@@ -1,5 +1,16 @@
-let newPostButton = document.querySelector(".logout")
-newPostButton.onclick = () => {
-  localStorage.clear()
-  window.location = window.location.origin + "/"
+let logoutButton = document.querySelector(".logout")
+
+if(!localStorage.getItem('registerId') || !localStorage.getItem('userId')) {
+  logoutButton.innerText = "Sign Up/Log In"
+} else {
+  logoutButton.innerText = "Log Out"
+}
+
+logoutButton.onclick = () => {
+  if(!localStorage.getItem('registerId') || !localStorage.getItem('userId')) {
+    window.location = window.location.origin + "/signup"
+  } else {
+    localStorage.clear()
+    window.location = window.location.origin + "/login"
+  }
 }
